peer listener housed in presentation
