+++ src/services/cluster/actors/mod.rs
@@ -1,2 +1,1 @@
 pub(crate) mod actor;
-pub(crate) mod listening_actor;
