+++ src/services/cluster/peers/peer.rs
@@ -1,8 +1,7 @@
 use std::ops::Deref;
 use std::ops::DerefMut;
 
-use crate::services::cluster::actors::listening_actor::ListeningActorKillTrigger;
-use crate::services::cluster::actors::listening_actor::PeerListeningActor;
+use crate::presentation::cluster_in::peer_listeners::peer_listener::PeerListeningActor;
 use crate::services::cluster::command::cluster_command::ClusterCommand;
 use crate::services::cluster::peers::connected_types::WriteConnected;
 use crate::services::cluster::peers::identifier::PeerIdentifier;
@@ -10,7 +9,10 @@ use crate::services::cluster::peers::kind::PeerKind;
 use tokio::net::tcp::OwnedWriteHalf;
 use tokio::net::TcpStream;
 use tokio::sync::mpsc::Sender;
+use tokio::task::JoinHandle;
 use tokio::time::Instant;
+
+use super::connected_types::ReadConnected;
 
 #[derive(Debug)]
 pub(crate) struct Peer {
@@ -56,5 +58,19 @@ impl Deref for Peer {
 impl DerefMut for Peer {
     fn deref_mut(&mut self) -> &mut Self::Target {
         &mut self.w_conn.stream
+    }
+}
+
+pub(super) type KillTrigger = tokio::sync::oneshot::Sender<()>;
+
+#[derive(Debug)]
+pub(crate) struct ListeningActorKillTrigger(KillTrigger, JoinHandle<ReadConnected>);
+impl ListeningActorKillTrigger {
+    pub(crate) fn new(kill_trigger: KillTrigger, listning_task: JoinHandle<ReadConnected>) -> Self {
+        Self(kill_trigger, listning_task)
+    }
+    pub(crate) async fn kill(self) -> ReadConnected {
+        let _ = self.0.send(());
+        self.1.await.unwrap()
     }
 }
